requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
urllib3==2.1.0
gunicorn==21.2.0
